(multicore-node) Node runtime from calvin codegen branch, with slight modifications and in CC instead of C. Starting support for torch tensors in multicore-node code-gen.

+++ platform-node/src/main/resources/lib/art-node/include/actors-parser.h
@@ -53,6 +53,6 @@ void parseInteractively(void);
 /**
  * Spawn command server at the given port.
  */
-void spawnServer(unsigned int port);
+void spawnServer(short port);
 
 #endif /* PARSER_INCLUSION_GUARD */
